Added settings and credits menus to main menu UI
- Created SettingsMenu and CreditsMenu panels with Back buttons
- Renamed MainMenu script to MenuManager.
- Added menu navigation methods to MenuManager script
- Ensured proper panel activation/deactivation for menu transitions
