+++ Assets/Sprites/Items/Lever/Lever.png.meta
@@ -134,7 +134,8 @@ TextureImporter:
     secondaryTextures: []
     spriteCustomMetadata:
       entries: []
-    nameFileIdTable: {}
+    nameFileIdTable:
+      Lever_0: -4454914845596493489
   mipmapLimitGroupName: 
   pSDRemoveMatte: 0
   userData:
