+++ Assets/Sprites/Items/Lever/LeverPullPart.png.meta
@@ -134,7 +134,8 @@ TextureImporter:
     secondaryTextures: []
     spriteCustomMetadata:
       entries: []
-    nameFileIdTable: {}
+    nameFileIdTable:
+      LeverPullPart_0: -2640083611626149161
   mipmapLimitGroupName: 
   pSDRemoveMatte: 0
   userData:
