+++ Assets/Sprites/Player/Dash.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0eea878b918604f459141e489fb4e8e3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
